libflux: use size_t for message/blob sizes
Problem: some libflux interfaces use the C int type to represent
buffer and I/O sizes, but size_t, as an unsigned type with more
range, is more appropriate in public interfaces.

Convert the following public API interfaces to use size_t:

  flux_msg_get_payload()
  flux_msg_set_payload()
  flux_request_decode_raw()
  flux_request_encode_raw()
  flux_response_decode_raw()
  flux_response_encode_raw()
  flux_respond_raw()
  flux_event_encode_raw()
  flux_event_decode_raw()
  flux_event_publish_raw()
  flux_rpc_raw()
  flux_rpc_get_raw()
  flux_kvs_lookup_get_raw()
  flux_kvs_txn_put_raw()

  Update tests and in-tree users, where appropriate.

+++ src/broker/trace.c
@@ -54,7 +54,7 @@ static void trace_msg (flux_t *h,
     int type = 0;
     char buf[64];
     const char *topic = NULL;
-    int payload_size = 0;
+    size_t payload_size = 0;
     json_t *payload_json = NULL;
     int errnum = 0;
     const char *errstr = NULL;
