+++ src/common/libcontent/content.h
@@ -21,7 +21,7 @@ enum {
  */
 flux_future_t *content_load_byhash (flux_t *h,
                                     const void *hash,
-                                    int hash_len,
+                                    size_t hash_len,
                                     int flags);
 flux_future_t *content_load_byblobref (flux_t *h,
                                        const char *blobref,
@@ -32,17 +32,22 @@ flux_future_t *content_load_byblobref (flux_t *h,
  * Storage for 'buf' belongs to 'f' and is valid until 'f' is destroyed.
  * Returns 0 on success, -1 on failure with errno set.
  */
-int content_load_get (flux_future_t *f, const void **buf, int *len);
+int content_load_get (flux_future_t *f, const void **buf, size_t *len);
 
 /* Send request to store blob.
  */
-flux_future_t *content_store (flux_t *h, const void *buf, int len, int flags);
+flux_future_t *content_store (flux_t *h,
+                              const void *buf,
+                              size_t len,
+                              int flags);
 
 /* Get result of store request (hash or blobref).
  * Storage belongs to 'f' and is valid until 'f' is destroyed.
  * Returns 0 on success, -1 on failure with errno set.
  */
-int content_store_get_hash (flux_future_t *f, const void **hash, int *hash_len);
+int content_store_get_hash (flux_future_t *f,
+                            const void **hash,
+                            size_t *hash_len);
 int content_store_get_blobref (flux_future_t *f,
                                const char *hash_name,
                                const char **blobref);
