+++ src/common/libflux/rpc.h
@@ -43,7 +43,7 @@ flux_future_t *flux_rpc_vpack (flux_t *h,
 flux_future_t *flux_rpc_raw (flux_t *h,
                              const char *topic,
                              const void *data,
-                             int len,
+                             size_t len,
                              uint32_t nodeid,
                              int flags);
 
@@ -56,7 +56,7 @@ int flux_rpc_get (flux_future_t *f, const char **s);
 
 int flux_rpc_get_unpack (flux_future_t *f, const char *fmt, ...);
 
-int flux_rpc_get_raw (flux_future_t *f, const void **data, int *len);
+int flux_rpc_get_raw (flux_future_t *f, const void **data, size_t *len);
 
 /* Accessor for RPC matchtag (see RFC 6).
  */
